Made modifications
Modified Board.java: Adding a private field to keep track of whose turn it is currently. Implemented the makeMove method but it does check if moves are legal. Added two helper methods to convert the string representation of a string and column to an int. Added a move method to Player.java.