Gradle
These are the files for the gradle build. It is not set up since i did not want to mess up the layout of our repository. The src directory does not ultimately contain any java files, since they are already in the board directory. The gradle build still has to be built, so the jar file hasn't been created yet.